Data Structure Generation
The game now recognizes the parent hero and enemy objects in field and
adds them to their respective 'field' lists.
The game also grabs from these lists the agility of the instances inside
and creates a map. This will determine turn order unless a better method
is found.

-Added agility stat
-Added turn-order map.

next up will be sorting the map on agility and being able to interact
using it.