RedoingThings
Hit a wall and lost momentum. Should learn to better document changes.
starting the process of making the turnlist and combat in general
independent from specific jobber or hero lists. mainly because the turn
checking and enemy selection broke when dealing with enemy death,
attempted (and failed) to remedy it with internal conditions according
to the type of list to remove that entity altogether. now seeing if this
way works better.