Autoattack Finished
Autoattack is currently random. needs tweaks to make it explicit. next
on should be player death management.
Added sublimetext project file, I absolutely hate gm's default editor so
this is a good change of pace.

What is currently needed is a GUI element for each player instance, and
death handling for players.